Apply proxy settings on AppRunners' Connection too
Proxy settings were applied only on WebEngine and the WebKit  window,
but not for network communication of the AppRunner process. As a result,
Last.fm scrobbling didn't work with manual proxy settings.

Proxy settings are now stored and handled in Connection instead of
WebEngine and applied also on Soup Session of the AppRunner process.

Issue: tiliado/nuvolaplayer#168

Signed-off-by: Jiří Janoušek <janousek.jiri@gmail.com>